+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Add support for `transaction.submit_for_partial_settlement!` to gateway instance
 * Add support for `transaction.void!` to gateway instance
 * Add support for `profile_id` in Transaction#create options for VenmoAccounts
+* Add support for `association_filter_id` in Customer#find
 
 # 2.83.0
 * Add support for `tax_amount` field on transaction `line_items`